+++ Changelog.md
@@ -1,5 +1,6 @@
 --------------------------Changelog for v1.9.9-beta4 of SecurityCraft--------------------------
 
+- New: Block and item tags "securitycraft:reinforced/concrete" and "securitycraft:reinforced/glazed_terracotta"
 - Fix: Compatibility with NeoForge 21.0.10-beta and newer
 - Misc.: The minimum required NeoForge version is now 21.0.10-beta